+++ Changelog.md
@@ -4,6 +4,7 @@
 * CXV1-9005 - Created split-button with submenu for use in toolbar view of email
 * CXV1-9007 - Hide interaction bar based on number of interactions
 * CXV1-9014 - Added side panel styling, hide side panel when no script, and added button to open side panel when collapsed in toolbarMode
+* CXV1-9013 - Toolbar view of custom fields
 
 ### [1.1.0] (2017-07-21)
 * BUGFIX - SDK version bump to 5.3.27-SNAPSHOT